Функции из client.py и server.py вынесены в отдельный файл.
