+++ .gitignore
@@ -35,6 +35,9 @@ MANIFEST
 pip-log.txt
 pip-delete-this-directory.txt
 
+# Client log
+log.txt
+
 # Unit test / coverage reports
 htmlcov/
 .tox/
