--- log.txt
@@ -1,5 +1,0 @@
-Лог создан: 2019-09-27 09:45:18.291144. Отправлено сообщение shjddje длиной 7 символов. Порт 9091, хост localhost
-Лог создан: 2019-09-27 09:45:19.150317. Отправлено сообщение efjedf длиной 6 символов. Порт 9091, хост localhost
-Лог создан: 2019-09-27 09:45:20.009468. Отправлено сообщение eflefd длиной 6 символов. Порт 9091, хост localhost
-Лог создан: 2019-09-27 09:46:59.656984. Отправлено сообщение sdhedh длиной 6 символов. Порт 9091, хост localhost
-Лог создан: 2019-09-27 09:47:00.719177. Отправлено сообщение ejffje длиной 6 символов. Порт 9091, хост localhost
